fix(data/repos): set spatial column type as geometry(Point,4326) and keep GiST index; remove HasSrid dependency;align repos with long IDs across APIs; update tuple types; make Spawn.CreateAsync accept long? cardId; ensure NearbyAsync/ClaimAsync/CatchAsync signatures; minor cleanups

+++ NeuroNexusBackend/NeuroNexusBackend/DTOs/SpawnCreateRequestDTO.cs
@@ -17,7 +17,7 @@ namespace NeuroNexusBackend.DTOs
         public double Lon { get; set; }
 
         /// <summary>Optional card to preview/grant at this spawn.</summary>
-        [Range(1, int.MaxValue)]
+        [Range(1, long.MaxValue)]
         public long? CardId { get; set; }
 
         /// <summary>Optional expiration timestamp (UTC).</summary>
